learning_service/pyproject.toml
[project]
name = "learning-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.41",
]
